Refactor: Relocate Tensor protobuf to tsercom/tensor/proto/
This change moves the `Tensor` protobuf message definition from its original
location in `tsercom/rpc/proto/common.proto` to a dedicated file at
`tsercom/tensor/proto/tensor.proto`.

The refactoring involved the following key steps:

1.  **Initial Relocation (Attempt 1):**
    *   I moved the `Tensor` message to `tsercom/tensor/rpc/proto/tensor.proto`.
    *   I updated `scripts/generate_protos.py` to handle this new file location.
    *   I fixed Python `import` statements across the `tsercom` codebase to reflect this intermediate path.

2.  **Path Correction (Your Feedback):**
    *   Based on your feedback, I revised the target path for `Tensor` to the simpler `tsercom/tensor/proto/tensor.proto`.

3.  **Revised Relocation (Attempt 2):**
    *   I moved `tensor.proto` from `tsercom/tensor/rpc/proto/` to `tsercom/tensor/proto/`.
    *   I updated the `package` declaration in `tensor.proto` to `tsercom.tensor.proto;`.
    *   I cleaned up the empty `tsercom/tensor/rpc/proto/` directory.
    *   I updated `scripts/generate_protos.py` again for the new `tensor/proto/tensor.proto` path and adjusted import paths for generation.
    *   I re-ran `scripts/generate_protos.py` successfully.
    *   I systematically updated Python import statements throughout `tsercom/` to use the final path (e.g., `from tsercom.tensor.proto import Tensor` or `from tsercom.tensor.proto.tensor_pb2 import Tensor`). This involved both automated script fixes and manual corrections for specific cases.

4.  **Verification:**
    *   The `Tensor` message definition itself was not altered.
    *   Static Analysis Gate:
        *   `black`, `ruff`, and `mypy` checks passed after necessary corrections (e.g., `noqa` additions for generated code, import fixes).
        *   `pylint` score improved to 9.84/10, with all Error-level messages resolved. Remaining Pylint items are non-Error categories or relate to Pylint's handling of an ignored file (`_version.py`).
    *   Test Suite: The full test suite (`pytest --timeout=120`) was run twice and passed successfully after all changes.

This refactoring provides a dedicated and more logical location for the `Tensor` protobuf definition, improving code organization.
