+++ scripts/generate_protos.py
@@ -395,6 +395,11 @@ def generate_protos(project_root: Path) -> None:
         ["caller_id/proto", "rpc/proto", "timesync/common/proto"],
     )
     generate_proto_file(
+        package_dir,
+        "tensor/proto/tensor.proto",
+        ["tensor/proto", "timesync/common/proto"],
+    )
+    generate_proto_file(
         package_dir, "test/proto/e2e_test_service.proto", ["test/proto"]
     )
