+++ tsercom/rpc/proto/__init__.py
@@ -25,73 +25,72 @@ if not TYPE_CHECKING:
         from tsercom.rpc.proto.generated.v1_73.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
         )
     elif version_string == "v1_72":
         from tsercom.rpc.proto.generated.v1_72.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_71":
         from tsercom.rpc.proto.generated.v1_71.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_70":
         from tsercom.rpc.proto.generated.v1_70.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_69":
         from tsercom.rpc.proto.generated.v1_69.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_68":
         from tsercom.rpc.proto.generated.v1_68.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_67":
         from tsercom.rpc.proto.generated.v1_67.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_66":
         from tsercom.rpc.proto.generated.v1_66.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_65":
         from tsercom.rpc.proto.generated.v1_65.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_64":
         from tsercom.rpc.proto.generated.v1_64.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_63":
         from tsercom.rpc.proto.generated.v1_63.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     elif version_string == "v1_62":
         from tsercom.rpc.proto.generated.v1_62.common_pb2 import (
             TestConnectionCall,
             TestConnectionResponse,
-            Tensor,
+            Tensor,  # noqa: F401
         )
     else:
         # The 'name' variable for the error message is 'common'
@@ -107,4 +106,3 @@ else:  # When TYPE_CHECKING
     from tsercom.rpc.proto.generated.v1_73.common_pb2 import (
         TestConnectionResponse as TestConnectionResponse,
     )
-    from tsercom.rpc.proto.generated.v1_73.common_pb2 import Tensor as Tensor
