+++ tsercom/rpc/serialization/serializable_tensor.py
@@ -10,7 +10,7 @@ from typing import Optional
 
 import torch
 
-from tsercom.rpc.proto import Tensor as GrpcTensor
+from tsercom.tensor.proto import Tensor as GrpcTensor
 from tsercom.timesync.common.synchronized_timestamp import (
     SynchronizedTimestamp,
 )
